Reorganized cfg package and added some documentation
