+++ src/lermitage/intellij/extra/icons/cfg/SettingsService.java
@@ -7,8 +7,8 @@ import com.intellij.openapi.project.Project;
 import lermitage.intellij.extra.icons.ExtraIconProvider;
 import lermitage.intellij.extra.icons.Globals;
 import lermitage.intellij.extra.icons.Model;
-import lermitage.intellij.extra.icons.cfg.settings.SettingsIDEService;
-import lermitage.intellij.extra.icons.cfg.settings.SettingsProjectService;
+import lermitage.intellij.extra.icons.cfg.services.impl.SettingsIDEService;
+import lermitage.intellij.extra.icons.cfg.services.impl.SettingsProjectService;
 import lermitage.intellij.extra.icons.providers.Angular2IconProvider;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
