+++ src/lermitage/intellij/extra/icons/cfg/dialogs/ModelConditionDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="lermitage.intellij.extra.icons.cfg.ModelConditionDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="lermitage.intellij.extra.icons.cfg.dialogs.ModelConditionDialog">
   <grid id="27dc6" binding="dialogPanel" layout-manager="GridLayoutManager" row-count="8" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
