+++ src/lermitage/intellij/extra/icons/cfg/models/UserIconsSettingsTableModel.java
@@ -1,13 +1,13 @@
-package lermitage.intellij.extra.icons.cfg;
+package lermitage.intellij.extra.icons.cfg.models;
 
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
 
 public class UserIconsSettingsTableModel extends DefaultTableModel {
 
-    static int ICON_ROW_NUMBER = 0;
-    static int ICON_ENABLED_ROW_NUMBER = 1;
-    static int ICON_LABEL_ROW_NUMBER = 2;
+    public static int ICON_ROW_NUMBER = 0;
+    public static int ICON_ENABLED_ROW_NUMBER = 1;
+    public static int ICON_LABEL_ROW_NUMBER = 2;
 
     /**
      * Table columns type.
@@ -20,7 +20,7 @@ public class UserIconsSettingsTableModel extends DefaultTableModel {
      */
     private final boolean[] canEdit = new boolean[]{false, true, false};
 
-    UserIconsSettingsTableModel() {
+    public UserIconsSettingsTableModel() {
         super(new Object[][]{}, new String[]{"", "", "Description"});
     }
